Demo map Created!
Created a simple 1v1v1v1 battle map.
Couldn't make things easy so had some fun creating a custom shader that takes in world position data and wraps this around a model. This allows for seamless Textures between models.
Created a couple materials with the new shader and added them to the battle map.
Created a simple placeholder spawning platform and added it  to the scene.