Fix lack of gravity when facing wall
Fix player not sliding down walls when they've jumped into them,
by wrapping the player's hitbox in another collider with a frictionless
physics material, as suggested by numerous people on the interwebs.

This seems to be one of the better solutions to this particular issue,
though I am baffled that this is a problem in the first place, in this
world-famous and widely used game engine.