Made playerInputs assigned in menu scene be the same as in FPSmode.
- Both InputManager and playerInputs dontDestroyOnLoad to ensure correct splitscreen
- Made playerFactory and main menu controller transfer inputs and add correct listeners
- Created PlayerInputManagerController to keep track of playerInputs and provide helper-functions
- Added more documentation + removed superfluous code and removed unnecessary Find functions

While this splitscreen implementation might seem a bit weird at first, it's
actually the only sensible way to meet our needs. This is due to some very
questionable splitScreen updating functions in Unity that only use the assigned
index of the player that joined last, instead of an actual count to determine
how many screens are needed. All relevant fields for this are at an inaccessiable
level. To make matters even worse, playerInputManager will always automatically
assign what it percieves as 'new' playerInputs, which include playerInputs
that are transferred from another scene with DontDestroyOnLoad.