Fix duplicate gun spawning on InitializeGun
GunFactory is a MonoBehaviour and is thus spawned in when one of its
static methods is called. In its Start method, it calls InitializeGun
to instantiate the gun it is supposed to build. This is *only* useful
in the GunTest scene, where we want a gun to show up automatically,
based on what the factory starts with. When we use InstantiateGun to
create a gun from other scenes, the Start method is also triggered and
creates a superfluous copy of the gun.

With this commit, the issue is avoided by initializing in InstantiateGun
and in Start *only when the scene in GunTest*.