Fix decals showing on tether plug
The layer check in DecalModifier didn't actually work, since the layer
property presents the layer index and not a bitmask. Altered the check
slightly to check bitmask agains bitmask.

This made it possible to remove the has-hitbox-and-connected-player
check as well.

Also added a new tag for objects that should not have decals on them.
Currently unused; left the check commented out with a matching TODO.