Remove superfluous revolver animation trigging
Animator was triggered from both the body and the animator script.
Removed it from the animator as removing it from the body made me unable
to shoot, likely because the animator has to call methods on the body class.

Did not fix the eternally frustrating revolver bug, but it seems to take
longer to trigger now???