Fix revolver + lazur/coil not firing
Outlining the issue in Norwegian:
- skyt nest siste skudd (her er ammo > 1)
- vent lenge nok til at animasjonen må være ferdig (mens anim kjører er ammo fortsatt > 1)
- prøv å skyte siste skudd (her er ammo == 1 og også <= 1)
--> da får du ikke skutt

This ammo <= 1 check did fix the animation issue we had with you being able to trigger fire animation just as the body started reloading, but removing this not-fix meant I had to fix it otherwise. Adjusted animation speed to actually correspond to the animation length, and gave the laser barrel an actual reduction in fire rate, and now things seem to work smoothly.